Scenery3D: Allow landscape polygon overdrawn onto the 3D scenery. - Important to show (not simulated) effects of Earth curvature.

+++ plugins/Scenery3d/src/gui/Scenery3dDialog.hpp
@@ -54,7 +54,7 @@ private slots:
 
 	void updateTorchStrength(float val);
 	void updateTorchRange(float val);
-	void updateLazyDrawingInterval(float val);
+	void updateLazyDrawingInterval(double val);
 	void updateShadowFilterQuality(S3DEnum::ShadowFilterQuality quality);
 	void updateSecondDominantFaceEnabled();
